+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.3.0] - 2022-01-12 (DEV-BETA)
+## [1.3.1] - 2022-01-13 (DEV-BETA)
 ### Added
 - Initial public re-release on Moodle.org and GitHub by TNG Consulting Inc.
 - Fix-8: Default email subject is now customizable.
 - Fix-2: Invitation text is now editable in language customization editor.
 - Ability to customize the default email subject line.
-- Many new variables usable in the email template.
+- Many new variables now available in the email template.
+- Anonymous users (no account) can reject invitation to users without an account. Users with an account will need to login.
+- You can only accept invitations if you are logged in with the invited user account. You can no longer accept invitations for others.
+- Revoking an invitation is now indicated in the status column of the History. The date revoked is the date in the Expiration Date column.
 - CONTRIBUTING.md.
 - CHANGELOG.md.
 - composer.json
@@ -24,5 +27,8 @@ All notable changes to this project will be documented in this file.
 - Multiple fixes for Multi-Language Moodle environments.
 - Improved accessibility and rtl language of invitation emails.
 - Plugin now compatible and tested with Moodle 3.9, 3.10 and 3.11.
+- Refactored code.
+- FAQ in documentation.
+- Russian (ru) and Dutch (nl) language files are no longer included. They will only be available through Moodle language packs.
 - Updated documentation for customizing email messages.
 - Updated copyright notice for 2022.